refactor: extract FieldHelp and MenuItem settings sub-components
