+++ src/components/features/settings/llm-profiles/profile-actions-menu.tsx
@@ -11,54 +11,10 @@ import { TextCursor, Copy } from "lucide-react";
 import { cn } from "#/utils/utils";
 import { dropdownMenuListClassName } from "#/utils/dropdown-classes";
 import { I18nKey } from "#/i18n/declaration";
-import { ConversationNameContextMenuIconText } from "#/components/features/conversation/conversation-name-context-menu-icon-text";
 import EditIcon from "#/icons/u-edit.svg?react";
 import CheckCircleIcon from "#/icons/u-check-circle.svg?react";
 import DeleteIcon from "#/icons/u-delete.svg?react";
-
-interface MenuItemProps {
-  index: number;
-  icon: React.ReactNode;
-  label: string;
-  onClick: () => void;
-  onKeyDown: (e: React.KeyboardEvent, index: number) => void;
-  menuItemsRef: React.MutableRefObject<(HTMLButtonElement | null)[]>;
-  disabled?: boolean;
-  testId: string;
-}
-
-function MenuItem({
-  index,
-  icon,
-  label,
-  onClick,
-  onKeyDown,
-  menuItemsRef,
-  disabled,
-  testId,
-}: MenuItemProps) {
-  return (
-    <button
-      ref={(el) => {
-        // eslint-disable-next-line no-param-reassign
-        menuItemsRef.current[index] = el;
-      }}
-      type="button"
-      onClick={onClick}
-      onKeyDown={(e) => onKeyDown(e, index)}
-      disabled={disabled}
-      className={cn(
-        "group w-full cursor-pointer rounded px-2 py-2 text-start text-nowrap text-sm font-normal",
-        "text-[var(--oh-foreground)] hover:bg-[var(--oh-interactive-hover)]",
-        "disabled:cursor-not-allowed disabled:opacity-50 disabled:hover:bg-transparent",
-      )}
-      role="menuitem"
-      data-testid={testId}
-    >
-      <ConversationNameContextMenuIconText icon={icon} text={label} />
-    </button>
-  );
-}
+import { MenuItem } from "./profile-actions-menu-item";
 
 interface ProfileActionsMenuProps {
   onEdit: () => void;
